created camera component
